requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
flask-bcrypt==1.0.1
flask-cors==4.0.0
pdfplumber==0.10.3
requests==2.31.0
python-dotenv==1.0.0
werkzeug==3.0.1
wtforms==3.1.2 